+++ airbyte-local-cli-nodejs/src/constants/constants.ts
@@ -1,0 +1,12 @@
+// Intermediate file path
+export const FILENAME_PREFIX = 'faros_airbyte_cli';
+export const CONFIG_FILE = `${FILENAME_PREFIX}_config.json`;
+export const SRC_CONFIG_FILENAME = `${FILENAME_PREFIX}_src_config.json`;
+export const DST_CONFIG_FILENAME = `${FILENAME_PREFIX}_dst_config.json`;
+export const SRC_CATALOG_FILENAME = `${FILENAME_PREFIX}_src_catalog.json`;
+export const DST_CATALOG_FILENAME = `${FILENAME_PREFIX}_dst_catalog.json`;
+export const DEFAULT_STATE_FILE = 'state.json';
+export const SRC_INPUT_DATA_FILE = `${FILENAME_PREFIX}_src_data`;
+export const SRC_OUTPUT_DATA_FILE = `${FILENAME_PREFIX}_src_output`;
+export const TMP_WIZARD_CONFIG_FILENAME = `tmp_wizard_config.json`;
+export const TMP_SPEC_CONFIG_FILENAME = `tmp_spec.json`;
