+++ airbyte-local-cli-nodejs/src/logger.ts
@@ -1,0 +1,5 @@
+import pino from 'pino';
+import pretty from 'pino-pretty';
+
+// Create a pino logger instance
+export const logger = pino(pretty({colorize: true}));
